Added: Ball and Bonus Mutex Synchronization
Added a mutex to the functions that change the values of the ball and the bonus, because of the various threads.
Fixed a bug in the ball velocityRatio due to rounding/conversion errors.

+++ Arkanoid/DLL/GameStructs.h
@@ -21,7 +21,7 @@ typedef struct Position {
 typedef struct Ball {
 	int directionX;
 	int directionY;
-	double velocityRatio;
+	int velocityRatio;
 	Position position;
 	int playerIndex;
 	BOOL inPlay;
